+++ controllers/tests/index.js
@@ -1,8 +1,7 @@
-const practices=require('./qa');
-const theories=require('./theory');
+const practices = require("./qa-practices");
+const theories = require("./qa-theories");
 
-
-module.exports={
-    practices,
-    theories,
-}
+module.exports = {
+  practices,
+  theories,
+};
