+++ routes/api/tests.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const { tests: testsCtrl } = require("../../controllers");
 
-router.get("/qa", testsCtrl.practices.getPractice);
-router.get("/theories", testsCtrl.theories.getTheories);
+router.get("/qa-practices", testsCtrl.practices.getPractice);
+router.get("/qa-theories", testsCtrl.theories.getTheories);
 
 module.exports = router;
